+++ js/app.config.js
@@ -25,8 +25,8 @@ app.config = {
 		},
 		training: {
 			base_url		: 'http://app.dev.withease.io:3001/', //local url
-			api_url			: 'https://webcenter.sisyphus-industries.com/',
-			web_url			: 'https://webcenter.sisyphus-industries.com/',
+			api_url			: 'http://dev.webcenter.sisyphus-industries.com/',
+			web_url			: 'http://dev.webcenter.sisyphus-industries.com/',
 			sisbot_url  : 'http://192.168.1.4:3002', //talking to sisbot    //  ... or just put your URL in here '192.168.XX.XXX:3002' << for local Dev Env --insert your ip address + 3000
 			port		: 3001,
 			show_tg		: true
@@ -126,7 +126,7 @@ if (window.location.href.indexOf('localhost') > -1) {
 	if (env == 'beta' || !app.config.envs[env]) app.config.env = 'beta';
 	console.log("Env:", env);
 }
-if (window.location.href.indexOf('.local') > -1)        	app.config.env = 'sisbot';
+if (window.location.href.indexOf('.local') > -1)        app.config.env = 'sisbot';
 if (window.location.href.indexOf('192.168') > -1) 			app.config.env = 'sisbot';
 
 // for any url not including dev, assumes prod env